Move ssh private key file to base config.
Used by both uploader and testing services.
