+++ test/unit/services/testing/config_test.py
@@ -1,6 +1,3 @@
-from pytest import raises
-
-from mash.mash_exceptions import MashConfigException
 from mash.services.testing.config import TestingConfig
 
 
@@ -12,12 +9,3 @@ class TestTestingConfig(object):
     def test_get_log_file(self):
         assert self.empty_config.get_log_file('testing') == \
             '/var/log/mash/testing_service.log'
-
-    def test_get_ssh_private_key_file(self):
-        assert self.config.get_ssh_private_key_file() == \
-            '/etc/mash/testing_key'
-
-        with raises(MashConfigException) as error:
-            self.empty_config.get_ssh_private_key_file()
-
-        assert str(error.value)
